+++ apps/e2e/indicators-storybook-e2e/src/e2e/status-indicator.component.cy.ts
@@ -23,5 +23,5 @@ describe('indicators-storybook', () => {
         );
       });
     });
-  });
+  }, true);
 });