+++ libs/components/colorpicker/testing/src/colorpicker/colorpicker-dropdown-harness.filters.ts
@@ -1,0 +1,8 @@
+import { BaseHarnessFilters } from '@angular/cdk/testing';
+
+/**
+ * A set of criteria that can be used to filter a list of `SkyColorpickerDropdownHarness` instances.
+ */
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export interface SkyColorpickerDropdownHarnessFilters
+  extends BaseHarnessFilters {}