+++ libs/components/colorpicker/testing/src/colorpicker/colorpicker-harness.filters.ts
@@ -1,0 +1,7 @@
+import { SkyHarnessFilters } from '@skyux/core/testing';
+
+/**
+ * A set of criteria that can be used to filter a list of `SkyColorpickerHarness` instances.
+ */
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export interface SkyColorpickerHarnessFilters extends SkyHarnessFilters {}